refactor: backend restructure
